[frontend] Runtime-registered hints for CircuitBuilder (#1453)
## Summary

Enables external crates to define and register their own `Hint`
implementations with `CircuitBuilder`, and collapses the four hardcoded
hint opcodes (BigUintDivide, BigUintModPow, ModInverse,
Secp256k1Endosplit) onto the new generic path. Addresses BINIUS-30.

Five incremental commits (bottom to top):

1. **Deduplicate hint gate modules** — delete four byte-identical
`gate/*_hint.rs` files whose `emit_eval_bytecode` bodies were all the
same; route through a single helper and have `Opcode::shape` delegate to
the `Hint` trait's own `shape()`.
2. **Give hints stable, name-derived HintIds** — add `const NAME:
&'static str` to the `Hint` trait; derive `HintId` from `DefaultHasher`
over the name; make `HintRegistry::register::<T>()` idempotent so
repeated registrations of the same hint type share one entry (previously
each gate got its own id).
3. **Add CircuitBuilder::call_hint** — new public method `call_hint<T:
Hint>(hint, dims, inputs) -> Vec<Wire>`. Thread `HintRegistry` into
`GateData::shape` via a new `Opcode::Hint` variant that looks up its
shape by hint id.
4. **Collapse four hint opcodes into Opcode::Hint** — drop the dedicated
variants; rewrite the four public convenience methods
(`biguint_divide_hint`, etc.) as thin wrappers over `call_hint` (public
API preserved).
5. **Thread HintRegistry into const-prop** — plumb `&HintRegistry`
through `constant_propagation` so hint gates with all-constant inputs
can be folded just like any other deterministic gate.

## Test plan

- `cargo test -p binius-frontend` (147 tests, includes new
`test_call_hint_user_registered` and revived
`test_constant_propagation_with_hint`)
- `cargo test -p binius-circuits` (340 tests — broad coverage of the
four built-in hints through bignum / ECDSA circuits via the new
`call_hint` path)
- `cargo test -p binius-prover --test prove_verify` (ZK and non-ZK
SHA256 round-trips)

+++ crates/frontend/src/compiler/eval_form/interpreter.rs
@@ -479,7 +479,7 @@ impl<'a> Interpreter<'a> {
 
 	// Hint execution
 	fn exec_hint(&mut self, ctx: &mut ExecutionContext<'_>) {
-		let hint_id = self.read_u32() as usize;
+		let hint_id = self.read_u32();
 
 		// Read dimensions
 		let n_dimensions = self.read_u16() as usize;
